Set default version to 3.8 and update yum repos
New version of glusterfs is out and packages have been removed from
the public glusterfs yum repository. We have to use the buildlogs
centos repo instead.

Remove auto-upgrade option which is useless if version is specified.
Moreover, you probably do not want to auto-upgrade glusterfs as it
necessitates a downtime.